Move outstanding protocols-->handlers dependencies to protocols
