+++ aries_vcx/src/handlers/mod.rs
@@ -1,21 +1,15 @@
-use futures::future::BoxFuture;
-
-use crate::error::prelude::*;
 use crate::handlers::connection::connection::ConnectionState;
-use crate::protocols::proof_presentation::verifier::state_machine::VerifierState;
-use crate::messages::a2a::A2AMessage;
 use crate::protocols::connection::invitee::state_machine::InviteeState;
 use crate::protocols::connection::inviter::state_machine::InviterState;
 use crate::protocols::issuance::holder::state_machine::HolderState;
 use crate::protocols::issuance::issuer::state_machine::IssuerState;
 use crate::protocols::proof_presentation::prover::state_machine::ProverState;
+use crate::protocols::proof_presentation::verifier::state_machine::VerifierState;
 
 pub mod connection;
 pub mod issuance;
 pub mod proof_presentation;
 pub mod out_of_band;
-
-pub type SendClosure = Box<dyn Fn(A2AMessage) -> BoxFuture<'static, VcxResult<()>> + Send + Sync>;
 
 impl From<ConnectionState> for u32 {
     fn from(state: ConnectionState) -> u32 {
