+++ aries_vcx/src/protocols/connection/pairwise_info.rs
@@ -1,5 +1,4 @@
 use crate::error::VcxResult;
-use crate::handlers::connection::public_agent::PublicAgent;
 use crate::libindy::utils::signus::create_and_store_my_did;
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
@@ -21,11 +20,5 @@ impl PairwiseInfo {
     pub fn create() -> VcxResult<PairwiseInfo> {
         let (pw_did, pw_vk) = create_and_store_my_did(None, None)?;
         Ok(PairwiseInfo { pw_did, pw_vk })
-    }
-}
-
-impl From<&PublicAgent> for PairwiseInfo {
-    fn from(agent: &PublicAgent) -> Self {
-        agent.pairwise_info()
     }
 }
